# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ghost Coach API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /plan-items/{id}/action:
    post:
      operationId: planItemAction
      tags: [today]
      summary: Perform action on a plan item (done/partial/skip/start)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanItemActionInput"
      responses:
        "200":
          description: Updated plan item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanItem"
...
components:
  schemas:
...
    ActivityTemplate:
      type: object
      required: [id, userId, name, category, goalTags, minimumMinutes, preferredMinutes, maximumMinutes, intensity, requiresFloorSpace, requiresPool, requiresStairs, requiresShower, requiresEquipment, minimumPrivacy, publicSuitability, weatherMode, allowedLocationTypes, preferenceScore, active, isPhysioApproved, version, createdAt, updatedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        category:
          type: string
        goalTags:
          type: array
          items:
            type: string
        minimumMinutes:
          type: integer
        preferredMinutes:
          type: integer
        maximumMinutes:
          type: integer
        intensity:
          type: string
        requiresFloorSpace:
          type: boolean
        requiresPool:
          type: boolean
        requiresStairs:
          type: boolean
        requiresShower:
          type: boolean
        requiresEquipment:
          type: array
          items:
            type: string
        minimumPrivacy:
          type: string
        publicSuitability:
          type: string
        weatherMode:
          type: string
        allowedLocationTypes:
          type: array
          items:
            type: string
        preferenceScore:
          type: integer
        active:
          type: boolean
        isPhysioApproved:
          type: boolean
        instructionsMarkdown:
          type: ["string", "null"]
        version:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
...
    PlanItem:
      type: object
      required: [id, dailyPlanId, activityTemplateId, role, scheduledStartAt, scheduledEndAt, state, createdAt, updatedAt]
      properties:
        id:
          type: string
        dailyPlanId:
          type: string
        activityTemplateId:
          type: string
        activityTemplate:
          $ref: "#/components/schemas/ActivityTemplate"
        role:
          type: string
        scheduledStartAt:
          type: string
        scheduledEndAt:
          type: string
        locationId:
          type: ["string", "null"]
        locationLabelSnapshot:
          type: ["string", "null"]
        candidateScore:
          type: ["number", "null"]
        selectionReason:
          type: ["string", "null"]
        state:
          type: string
        calendarEventId:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    PlanItemActionInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
        actualMinutes:
          type: integer
        skipReason:
          type: string
        notes:
          type: string
...